Removed winexts due to kallsyms_lookup_name being unexported.
This is not a real issue as we can still use pattern scanning
(for finding functions) but in order to make this a clean project,
winexts will be probably moved to another entire linux kernel modification.
This will make things much simpler in some ways, but I gotta maintain
the modifications and hopefully maybe accepted by the linux kernel
developers.
For now I'll focus more on developing the XLib project.